# Optimize motion receive