--- a/src/include/cdb/cdbinterconnect.h
+++ b/src/include/cdb/cdbinterconnect.h
@@ -475,8 +475,6 @@ typedef struct MotionNodeEntry
 		* value. */
 	uint64          sel_rd_wait;            /* Total time (usec) spent in select wait trying to read */
 	uint64          sel_wr_wait;            /* Total time spent (usec) in select wait trying to write */
-
-	uint64			memKB;	/* How much memory should this motion node use? */
 }       MotionNodeEntry;