--- a/src/include/cdb/cdbmotion.h
+++ b/src/include/cdb/cdbmotion.h
@@ -65,7 +65,7 @@ extern void InitMotionLayerNode(MotionLayerState *mlStates, int16 motNodeID);
 
 /* Initialization of each motion node in execution plan. */
 extern void UpdateMotionLayerNode(MotionLayerState *mlStates, int16 motNodeID, bool preserveOrder,
-								  TupleDesc tupDesc, uint64 operatorMemKB);
+								  TupleDesc tupDesc);
 
 /* Cleanup of each motion node in execution plan (normal termination). */
 extern void EndMotionLayerNode(MotionLayerState *mlStates, int16 motNodeID, bool flushCommLayer);