--- a/src/include/cdb/htupfifo.h
+++ b/src/include/cdb/htupfifo.h
@@ -46,29 +46,9 @@ typedef struct htup_fifo_state
 	htf_entry	freelist;
 	int			freelist_count;
 
-	/* A count of HeapTuples in the FIFO. */
-	int			tup_count;
-
-	/*
-	 * An estimate of the current in-memory size of the FIFO's contents, in
-	 * bytes.
-	 */
-	int			curr_mem_size;
-
-	/*
-	 * The maximum size that the FIFO is allowed to grow to, in bytes, before
-	 * it will cause an error to be reported.
-	 */
-	uint32		max_mem_size;
-
 }	htup_fifo_state, *htup_fifo;
 
-
-#define MIN_HTUPFIFO_KB 8
-
-
-extern htup_fifo htfifo_create(int max_mem_kb);
-extern void htfifo_init(htup_fifo htf, int max_mem_kb);
+extern htup_fifo htfifo_create(void);
 extern void htfifo_destroy(htup_fifo htf);
 
 extern void htfifo_addtuple(htup_fifo htf, GenericTuple htup);