This is some low-hanging fruit that came up, when I did some performance testing of Sorted Motion Receives. Some of these optimizations are not specific to Sorted Receives, though, and should help in many other scenarios, too. Some of these only have an effect in the `gp_enable_motion_mk_sort=off` codepath.

Why now? I was originally wondering why we have a GPDB-specific binary heap implementation, CdbHeap, when there's an almost identical one in upstream, in src/backend/lib/binaryheap.c. I'm not changing that as part of this patch, but that's what got me looking into this code.

Performance testing I did
====

Test setup:
```
create table sortedmotiontest (i int4) distributed by (i);
insert into sortedmotiontest select g from generate_series(1, 10000000) g;
create index on sortedmotiontest (i);
```

To test performance of Sorted Motion Receive, I used this query:
```
postgres=# explain select count(*) from (select i from sortedmotiontest order by i offset 1) x;
                                                            QUERY PLAN                                                            
----------------------------------------------------------------------------------------------------------------------------------
 Aggregate  (cost=1463652.05..1463652.06 rows=1 width=8)
   ->  Limit  (cost=0.32..1338652.06 rows=9999999 width=4)
         ->  Gather Motion 3:1  (slice1; segments: 3)  (cost=0.18..1338652.06 rows=10000000 width=4)
               Merge Key: sortedmotiontest.i
               ->  Index Only Scan using sortedmotiontest_i_idx on sortedmotiontest  (cost=0.18..1138652.06 rows=3333334 width=4)
 Planning time: 0.276 ms
 Optimizer: legacy query optimizer
(7 rows)
```

I repeated this query a few times, to not be fooled by random variance. I didn't bother to calculate an average, but the number shown here is a fairly representative one, on my laptop, with a three-segment cluster:

Before these patches:
```
postgres=# set gp_enable_motion_mk_sort=off;
SET
Time: 9.613 ms
postgres=# select count(*) from (select i from sortedmotiontest order by i offset 1) x;
  count  
---------
 9999999
(1 row)

Time: 3111.369 ms
```

After:
```
postgres=# set gp_enable_motion_mk_sort=off;
SET
Time: 0.788 ms
postgres=# select count(*) from (select i from sortedmotiontest order by i offset 1) x;
  count  
---------
 9999999
(1 row)

Time: 2437.183 ms
```


Profiling
-----

I took a profile of that with Linux perf (`perf report --call-graph-callee`). I expanded the pfree and AllocSetAllocImpl to show where those calls are coming from.
```
Samples: 92K of event 'cycles', Event count (approx.): 46659727978
  Children      Self  Command     Shared Object        Symbol                                                                                                                                ◆
+   70.48%     7.00%  postmaster  postgres             [.] ExecProcNode                                                                                                                      ▒
+   64.29%     0.79%  postmaster  postgres             [.] ExecAgg                                                                                                                           ▒
+   59.24%     3.31%  postmaster  postgres             [.] ExecMotion                                                                                                                        ▒
+   58.40%     1.16%  postmaster  postgres             [.] ExecLimit                                                                                                                         ▒
+   37.38%     5.59%  postmaster  postgres             [.] RecvTupleFrom                                                                                                                     ▒
+   28.04%     0.00%  postmaster  postgres             [.] standard_ExecutorRun                                                                                                              ▒
+   18.67%     3.78%  postmaster  postgres             [.] SiftDown                                                                                                                          ▒
+   18.42%     0.00%  postmaster  postgres             [.] PostgresMain                                                                                                                      ▒
+   18.42%     0.00%  postmaster  postgres             [.] exec_simple_query                                                                                                                 ▒
+   18.41%     0.00%  postmaster  postgres             [.] PortalRun                                                                                                                         ▒
+   18.41%     0.00%  postmaster  postgres             [.] PortalRunSelect                                                                                                                   ▒
+   16.74%     0.00%  postmaster  postgres             [.] ExecutePlan                                                                                                                       ▒
+   15.94%     4.74%  postmaster  postgres             [.] CdbMergeComparator                                                                                                                ▒
+   14.71%     2.45%  postmaster  postgres             [.] CvtChunksToTup                                                                                                                    ▒
-   10.57%     9.98%  postmaster  postgres             [.] AllocSetAllocImpl                                                                                                                 ▒
   - 7.18% palloc                                                                                                                                                                            ▒
      + 2.78% htfifo_addtuple                                                                                                                                                                ▒
      + 2.23% CvtChunksToTup                                                                                                                                                                 ▒
      + 2.17% initStringInfoOfSize                                                                                                                                                           ▒
   + 2.21% AllocSetAlloc                                                                                                                                                                     ▒
     0.63% AllocSetAllocImpl                                                                                                                                                                 ▒
-    9.37%     1.91%  postmaster  postgres             [.] pfree                                                                                                                             ▒
   - 9.24% pfree                                                                                                                                                                             ▒
      + 3.06% ExecStoreMinimalTuple                                                                                                                                                          ▒
      + 2.33% clearTCList                                                                                                                                                                    ▒
      + 1.95% CvtChunksToTup                                                                                                                                                                 ▒
      + 1.78% htfifo_gettuple                                                                                                                                                                ▒
+    8.63%     3.04%  postmaster  postgres             [.] advance_aggregates                                                                                                                ▒
+    8.29%     7.82%  postmaster  postgres             [.] memtuple_getattr_by_alignment                                                                                                     ▒
+    7.97%     0.73%  postmaster  postgres             [.] palloc                                                                                                                            ▒
+    7.89%     0.44%  postmaster  postgres             [.] memtuple_getattr                                                                                                                  ▒
+    7.80%     7.36%  postmaster  postgres             [.] AllocSetFreeImpl                                                                                                                  ▒
+    7.29%     2.66%  postmaster  postgres             [.] htfifo_gettuple                    
```

These patches address the overheads shown in the profile like this:

* The "Remove limit on htupfifo freelist size" commit removes the palloc/pfree overhead from htfifo_addtuple/getuple.

* The "Optimize CvtChunksToTup" commit eliminates the palloc/pfree overhead from initStringInfoOfSize.

* The "Avoid overhead of extracting first key column in Sorted Motion Receiver" commit reduces the number of memtuple_getattr calls.

* The "Use SortSupport in Motion node" reduces the time spent in CdbMergeComparator.


With these patches applied, the profile looks like this:
```
Samples: 71K of event 'cycles', Event count (approx.): 33545443345
  Children      Self  Command     Shared Object        Symbol                                                                                                                                ◆
+   66.91%     9.41%  postmaster  postgres             [.] ExecProcNode                                                                                                                      ▒
+   59.65%     1.16%  postmaster  postgres             [.] ExecAgg                                                                                                                           ▒
+   52.02%     7.23%  postmaster  postgres             [.] ExecMotion                                                                                                                        ▒
+   51.71%     1.70%  postmaster  postgres             [.] ExecLimit                                                                                                                         ▒
+   28.94%     7.43%  postmaster  postgres             [.] RecvTupleFrom                                                                                                                     ▒
+   18.19%     0.00%  postmaster  postgres             [.] standard_ExecutorRun                                                                                                              ▒
+   12.04%     3.96%  postmaster  postgres             [.] advance_aggregates                                                                                                                ▒
+   10.56%     6.38%  postmaster  postgres             [.] SiftDown                                                                                                                          ▒
+   10.38%     0.00%  postmaster  postgres             [.] ExecutePlan                                                                                                                       ▒
-    7.91%     7.27%  postmaster  postgres             [.] AllocSetAllocImpl                                                                                                                 ▒
   - 6.72% palloc                                                                                                                                                                            ▒
      + 3.38% CvtChunksToTup                                                                                                                                                                 ▒
      + 3.33% RecvTupleChunk                                                                                                                                                                 ▒
     0.67% AllocSetAllocImpl                                                                                                                                                                 ▒
-    7.74%     2.54%  postmaster  postgres             [.] pfree                                                                                                                             ▒
   - 7.54% pfree                                                                                                                                                                             ▒
      + 4.53% ExecStoreMinimalTuple                                                                                                                                                          ▒
      + 2.81% clearTCList                                                                                                                                                                    ▒
+    7.45%     0.65%  postmaster  postgres             [.] palloc                                                                                                                            ▒
+    6.51%     1.93%  postmaster  postgres             [.] CvtChunksToTup                                                                                                                    ▒
+    6.23%     1.57%  postmaster  postgres             [.] ExecStoreMinimalTuple                                                                                                             ▒
+    6.12%     0.15%  postmaster  postgres             [.] CdbHeap_DeleteMinAndInsert                                                                                                        ▒
+    5.99%     0.07%  postmaster  [kernel.kallsyms]    [k] entry_SYSCALL_64_after_swapgs                                                                                                     ▒
+    5.92%     0.04%  postmaster  [kernel.kallsyms]    [k] do_syscall_64                                                                                                                     ▒
+    5.63%     5.16%  postmaster  postgres             [.] AllocSetFreeImpl                                                                                                                  ▒
+    5.17%     4.76%  postmaster  postgres             [.] BackoffBackendTick                                                                                                                ▒
+    4.66%     0.97%  postmaster  postgres             [.] RecvTupleChunk                                                                                                                    ▒
+    4.59%     0.08%  postmaster  [kernel.kallsyms]    [k] udp_sendmsg                                                                                                                       ▒
+    4.54%     0.02%  postmaster  [kernel.kallsyms]    [k] ktime_get_ts64                                                                                                                    ▒
+    4.53%     0.02%  postmaster  [kernel.kallsyms]    [k] read_tsc                                                                                                                          ▒
+    4.40%     2.31%  postmaster  postgres             [.] ExecProject                                                                                                                       ▒
+    4.32%     1.24%  postmaster  postgres             [.] advance_transition_function                                                                                                       ▒
+    4.28%     0.09%  postmaster  [kernel.kallsyms]    [k] udpv6_recvmsg                                                                                                                     ▒
+    4.28%     0.05%  postmaster  postgres             [.] RecvTupleChunkFromUDPIFC                                                                                                          ▒
+    4.25%     0.01%  postmaster  [kernel.kallsyms]    [k] __intel_pmu_enable_all.constprop.17                                                                                               ▒
+    4.23%     3.09%  postmaster  postgres             [.] CdbMergeComparator                                                                                                                ▒
```